Removed HASHMAP macro.
git-svn-id: svn://svn.renegadeprojects.com/ares/trunk@52 859b54a9-7a54-0410-aeb3-f8d2f1fa40fd

+++ TechnoTypeExt.cpp
@@ -61,8 +61,8 @@ void TechnoTypeClassExt::TechnoTypeClassData::Initialize(TechnoTypeClass *pThis)
 	this->Survivors_Pilots.SetCapacity(SideClass::Array->get_Count(), NULL);
 
 	this->Survivors_PassengersEscape = pThis->get_Crewed();
-	this->Survivors_PilotChance = RulesClass::Global()->get_CrewEscape() * 100;
-	this->Survivors_PassengerChance = RulesClass::Global()->get_CrewEscape() * 100;
+	this->Survivors_PilotChance = (int)RulesClass::Global()->get_CrewEscape() * 100;
+	this->Survivors_PassengerChance = (int)RulesClass::Global()->get_CrewEscape() * 100;
 
 	for(int i = 0; i < SideClass::Array->get_Count(); ++i)
 	{
